Moved forgotten images to the img folder
